+++ src/controllers/page-controller.js
@@ -7,7 +7,7 @@ import SortComponent from "../components/sort-component";
 import StatisticComponent from "../components/statistic-component";
 import FilmsComponent from "../components/films-component";
 import FilmsController from "./films-controller";
-import FilmsModel from "../models/films";
+import FilmsModel from "../models/films-model";
 import FilterController from "./filter-controller";
 import FooterStatisticsComponent from "../components/footer-statistics-component";
 import FilmsNoData from "../components/films-no-data";
