+++ src/models/comment-model.js
@@ -1,4 +1,4 @@
-export default class Comment {
+export default class CommentModel {
   constructor(data) {
     this.id = data[`id`];
     this.text = data[`comment`];
@@ -16,10 +16,10 @@ export default class Comment {
   }
 
   static parseComment(data) {
-    return new Comment(data);
+    return new CommentModel(data);
   }
 
   static parseComments(data) {
-    return data.map(Comment.parseComment);
+    return data.map(CommentModel.parseComment);
   }
 }
